+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.1 - 18-05-2026
+
+### Added
+
+- `TelegramEx.Builder.Poll` for sending polls and quizzes
+- Proper library documentation
+
 ## [1.2.0] - 13-04-2026
 
 ### Added